+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 # Change Log
 
-All notable changes to the "vscode-emacs-indent" extension will be documented in this file.
+All notable changes to the "vscode-emacs-indent-v2" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
 
-- Initial release
+	- Initial release
+
+## [0.0.9] - 2023-04-30 - kshartman
+
+### Changed
+	
+	- Use formatDocument editor primitive instead of indentLine.
+	- If more than one line selected, indent the region
+